requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlmodel
asyncpg
firebase-admin
httpx
python-multipart
python-dotenv
openai
google-generativeai
pytest
